msm_flash: flashlight notif: reschedule if in long flashless wait
If phone was not able to flash last notif (not face down etc), the
alarm job went in a long wait mode for vibration notif reminder.
At next flash call, check this, and if so, reschedule the job, so
that it can start flashing if it's face down since then, no need to
wait half a minute, delaying the new notification's instant timing.

Signed-off-by: Eliminater74 <eliminater74@gmail.com>